Change unsigned int to size_t.
git-svn-id: https://subversion.assembla.com/svn/qmcdev/trunk@5070 e5b18d87-469d-4833-9cc0-8cdfa06e9491

+++ src/Estimators/EstimatorManager.h
@@ -292,7 +292,7 @@ private:
     void collectBlockAverages(int num_threads=1);
     ///add header to an ostream
     void addHeader(ostream& o);
-    unsigned long FieldWidth;
+    size_t FieldWidth;
   };
 }
 #endif
